Move the scalar function family out of externals.h
externals.h is the machinery for handle types that cross the boundary, but
ScalarFunctionHolder owns family-specific state -- the callbacks and the user's
extra info -- so it and everything behind it belong with the family, not in the
externals list.

The block moves as a unit because it is one dependency chain: the holder needs
the extra info complete, and CreateExternalForScalarFunction needs the holder
complete. Splitting it anywhere would trade the move for a declaration and
definition split.

Doing so also removes a split introduced earlier in this branch. The callback
traits had to be declared in externals.h and defined in scalar_function_helpers.h
purely because CreateExternalForVectorWithoutFinalizer is defined at the end of
externals.h, below where the traits were needed. Below the external helpers they
are ordinary inline structs, defined once.

The tradeoff is that the type tag list in externals.h is no longer complete.
Collecting all tags into a header of their own would restore that, and stays
available: tags are pure data, so nothing here blocks it.

Table functions should follow the same shape in their own header.
