+++ bindings/src/externals.h
@@ -2,10 +2,6 @@
 
 #include "napi_setup.h"
 #include "duckdb.h"
-#include "napi_ref_reaper.h"
-#include "duckdb_thread_callback.h"
-#include <cstddef>
-#include <memory>
 
 // Externals
 
@@ -325,97 +321,6 @@ inline Napi::External<duckdb_result> CreateExternalForResult(Napi::Env env, duck
 
 inline duckdb_result *GetResultFromExternal(Napi::Env env, Napi::Value value) {
   return GetDataFromExternal<duckdb_result>(env, ResultTypeTag, value, "Invalid result argument");
-}
-
-static const napi_type_tag ScalarFunctionTypeTag = {
-  0x95D48B7051D14994, 0x9F883D7DF5DEA86D
-};
-
-// Per-callback details for DuckDBThreadCallback. Defined in
-// scalar_function_helpers.h, which is included after the CreateExternalFor*
-// helpers these need.
-
-struct ScalarFunctionBindCallbackTraits {
-  using Payload = duckdb_bind_info;
-  static const char *ResourceName();
-  static void Call(Napi::Env env, Napi::Function callback, const Payload &payload);
-  static void SetError(const Payload &payload, const char *message);
-};
-
-struct ScalarFunctionMainCallbackTraits {
-  struct Payload {
-    duckdb_function_info info;
-    duckdb_data_chunk input;
-    duckdb_vector output;
-  };
-  static const char *ResourceName();
-  static void Call(Napi::Env env, Napi::Function callback, const Payload &payload);
-  static void SetError(const Payload &payload, const char *message);
-};
-
-struct ScalarFunctionInternalExtraInfo {
-  DuckDBThreadCallback<ScalarFunctionBindCallbackTraits> bind_callback;
-  DuckDBThreadCallback<ScalarFunctionMainCallbackTraits> main_callback;
-  std::shared_ptr<ManagedObjectReference> user_extra_info_ref;
-
-  explicit ScalarFunctionInternalExtraInfo(const std::shared_ptr<NapiRefReaper> &env_state)
-    : bind_callback(env_state), main_callback(env_state) {}
-
-  void SetBindFunction(Napi::Env env, Napi::Function func) {
-    bind_callback.Set(env, func);
-  }
-
-  void SetMainFunction(Napi::Env env, Napi::Function func) {
-    main_callback.Set(env, func);
-  }
-
-  void SetUserExtraInfo(const std::shared_ptr<NapiRefReaper> &reaper, Napi::Object user_extra_info) {
-    user_extra_info_ref = user_extra_info.IsUndefined() ? nullptr : MakeManagedObjectReference(reaper, user_extra_info);
-  }
-};
-
-inline void DeleteScalarFunctionInternalExtraInfo(ScalarFunctionInternalExtraInfo *internal_extra_info) {
-  delete internal_extra_info;
-}
-
-struct ScalarFunctionHolder {
-  duckdb_scalar_function scalar_function;
-  ScalarFunctionInternalExtraInfo *internal_extra_info;
-
-  ScalarFunctionHolder(duckdb_scalar_function scalar_function_in): scalar_function(scalar_function_in), internal_extra_info(nullptr) {}
-
-  ~ScalarFunctionHolder() {
-    // duckdb_destroy_scalar_function is a no-op if already destroyed
-    duckdb_destroy_scalar_function(&scalar_function);
-  }
-
-  ScalarFunctionInternalExtraInfo *EnsureInternalExtraInfo(const std::shared_ptr<NapiRefReaper> &env_state) {
-    if (!internal_extra_info) {
-      internal_extra_info = new ScalarFunctionInternalExtraInfo(env_state);
-      duckdb_scalar_function_set_extra_info(scalar_function, internal_extra_info, reinterpret_cast<duckdb_delete_callback_t>(DeleteScalarFunctionInternalExtraInfo));
-    }
-    return internal_extra_info;
-  }
-};
-
-inline ScalarFunctionHolder *CreateScalarFunctionHolder(duckdb_scalar_function scalar_function) {
-  return new ScalarFunctionHolder(scalar_function);
-}
-
-inline void FinalizeScalarFunctionHolder(Napi::BasicEnv, ScalarFunctionHolder *holder) {
-  delete holder;
-}
-
-inline Napi::External<ScalarFunctionHolder> CreateExternalForScalarFunction(Napi::Env env, duckdb_scalar_function scalar_function) {
-  return CreateExternal<ScalarFunctionHolder>(env, ScalarFunctionTypeTag, CreateScalarFunctionHolder(scalar_function), FinalizeScalarFunctionHolder);
-}
-
-inline ScalarFunctionHolder *GetScalarFunctionHolderFromExternal(Napi::Env env, Napi::Value value) {
-  return GetDataFromExternal<ScalarFunctionHolder>(env, ScalarFunctionTypeTag, value, "Invalid scalar function argument");
-}
-
-inline duckdb_scalar_function GetScalarFunctionFromExternal(Napi::Env env, Napi::Value value) {
-  return GetScalarFunctionHolderFromExternal(env, value)->scalar_function;
 }
 
 static const napi_type_tag ValueTypeTag = {
